# refactor : Calculator Class 를 index.js 에서 분리
